menambahkan fitur login, checkout, pesanan dan pembayaran

+++ database/migrations/2024_11_12_160635_create_products_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->string("title");
             $table->string("image");
             $table->bigInteger("price");
-            $table->integer("stok")->default(0);
-            $table->integer('qty_out')->default(0);
+            $table->float("stok")->default(0);
+            $table->float('qty_out')->default(0);
             $table->text("description");
             $table->timestamps();
         });
